Fix Require cycle
