+++ packages/client/src/controllers/publisher.ts
@@ -4,7 +4,7 @@ import { generateChildLogger, getLoggerContext } from "@walletconnect/logger";
 import { IRelayer, PublisherTypes, RelayerTypes } from "@walletconnect/types";
 
 import { PUBLISHER_CONTEXT, HEARTBEAT_EVENTS, PUBLISHER_DEFAULT_TTL } from "../constants";
-import { getRelayProtocolName, getRelayProtocolApi } from "./relayer";
+import { getRelayProtocolName, getRelayProtocolApi } from "./relayProtocol";
 import { IEvents, JsonRpcPayload, RequestArguments } from "@walletconnect/jsonrpc-types";
 import { RelayJsonRpc } from "@walletconnect/relay-api";
